refactor: replace fetch calls with api client
